🐛[Cable] - Fix the anchor impulsion by allowing smaller mass for attached node
Push the inverse mass limit to the DBL_EPSILON(2.2204460492503131e-016).

+++ src/BulletCable/btCable.cpp
@@ -178,7 +178,7 @@ void btCable::PrepareSolver()
 
 		// Node's masses
 		const double invMassNode = n->m_im;
-		const double massNode = invMassNode < FLT_EPSILON ? 0.0 : 1.0 / invMassNode;
+		const double massNode = invMassNode < DBL_EPSILON ? 0.0 : 1.0 / invMassNode;
 
 		// Body's masses
 		const double invMassBody = b->getInvMass();
